Wrote FileUtilities and updated the simulation
I rewrote the old config_scan, zmatrix_scan, and opls_scan code files
and merged them into FileUtilities.cpp. This is a monolithic file that
has everything. I did it like this in the interest of time, although I
would like to split it up a bit. I integrated the new FileUtilities with
the simulation, serial calcs, parallel calcs, serial box, box, and
parallel box. All of these were updated to handle filling the box with
data from the files.
I tested it out and the program compiles with metrosim.make correctly.
The program will run in serial (haven't tested parallel), although the
numbers are wrong. I am pushing so Seth and Josh can help debug the
issues.

+++ src/Metropolis/Simulation.h
@@ -25,7 +25,7 @@ class Simulation
 	private:
 		Box *box;
 		SimulationArgs args;
-		int simSteps;
+		long simSteps;
 };
 
 #endif
